Image Overhaul
- Move image handling to their own controller
- Handle all profile and banner pic uploading
- Make imageDelete call agnostic to handle all imageDeletion needs
