+++ index.js
@@ -7,7 +7,7 @@ const projectRoutes = require('./routes/project_routes');
 const communityRoutes = require('./routes/community_routes');
 const userRoutes = require('./routes/user_routes');
 const postRoutes = require('./routes/post_routes');
-const uploadRoutes = require('./routes/upload_routes');
+const imageRoutes = require('./routes/image_routes');
 const reactionRoutes = require('./routes/reaction_routes');
 const configRoutes = require('./routes/config_routes');
 const commentRoutes = require('./routes/comment_routes');
@@ -68,7 +68,7 @@ communityRoutes(app);
 userRoutes(app);
 postRoutes(app);
 reactionRoutes(app);
-uploadRoutes(app);
+imageRoutes(app);
 configRoutes(app);
 commentRoutes(app);
 upvoteRoutes(app);
